More efficient solution to the problem of eliminating LAMBDAS {} values, for which it is impossible to determine whether the set corresponds to escaping lambdas or non-escaping lambdas.
Rather than initializing all variables of function type to 
LAMDBAS {proxy}, we initialize all variables of function type to
LAMBDAS {}.  After running the analysis to a fixed-point, we "bump"
any variable of function type that is still at LAMBDAS {} to 
LAMBDAS {proxy} and run the analysis again to a fixed-point.  This
avoid creating too many proxy functions, which overwhelms the LAMBDAS
sets with proxies and slows down the analysis.



git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/trunk@1993 ac4fbcad-d21d-0410-b760-c7b1d3380187